+++ lib/components/App.js
@@ -1,133 +1,29 @@
 import React, { Component } from 'react';
-const fs = window.require('fs');
-const electron = window.require('electron');
-const remote = electron.remote;
-const ipc = electron.ipcRenderer;
-const mainProcess = remote.require('./main.js');
-import ImageCard from './ImageCard.jsx';
-import ReactPaint from 'react-paint';
-import FileSaver from 'FileSaver';
-
-const injectTapEventPlugin = require('react-tap-event-plugin');
-injectTapEventPlugin();
-
-import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
-import AutoComplete from 'material-ui/AutoComplete';
+import SaveScreenshot from './SaveScreenshot.jsx';
+import SubmitSuccess from './SubmitSuccess.jsx';
 
 export default class App extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      imgData: '',
-      directories: [],
-      folder: '',
-      newName: '',
-      errorMessage: false
+      submitted: false
     }
-
-    this.saveFile = this.saveFile.bind(this);
-    this.fileCheck = this.fileCheck.bind(this);
-    this.viewPhotos = this.viewPhotos.bind(this);
   }
 
-  componentDidMount() {
-    ipc.send('mounted', 'mounted')
-
-    ipc.on('currentImg', (event, imgData) => {
-      this.setState({ imgData })
-      this.setState({ directories: imgData.directories })
-    })
+  example(e) {
+    console.log(e);
   }
-
-  fileCheck(e) {
-    const { folder, imgData } = this.state;
-    const { file } = imgData;
-    const newName = e.target.value;
-
-    mainProcess.fileCheck({ folder, newName, file })
-
-    ipc.on('duplicate', (event, response) => {
-      if (response === true) {
-        this.setState({ errorMessage: true })
-      }
-      else if (response === false) {
-        this.setState({ errorMessage: false })
-      }
-    })
-  }
-
-  saveFile() {
-    console.log('save')
-    const { folder, newName, imgData } = this.state;
-    const { file } = imgData;
-
-    const fileToSave = {
-      folder,
-      newName,
-      file
-    }
-    mainProcess.saveFile(fileToSave);
-    this.setState({ folder: '', newName: '', imgData: '' })
-  }
-
-  viewPhotos() {
-      const { filePath } = this.state.imgData;
-      mainProcess.viewPhotos(filePath);
-    }
 
   render() {
-    const { imgData, folder, newName, directories, errorMessage } = this.state
-    const dispErrorMessage = () => {
-      if (errorMessage) {
-        console.log('errorMessage true')
-        return <p>This file already exists in this folder.  Please enter a new name</p>
-      }
+
+    const displayApp = () => {
+      return !this.state.submitted ?
+        <SaveScreenshot toggleView={() => this.setState({ submitted: true })} /> : <SubmitSuccess />
     }
 
     return (
       <div>
-        <div className='save-inputs'>
-          <button
-            className='view-btn'
-            onClick={this.viewPhotos}
-            >View My Photos
-          </button>
-          <label>
-            Desktop/snip-it-images/
-
-            <MuiThemeProvider>
-              <AutoComplete
-                hintText='Save to Folder'
-                onNewRequest={(folder) => this.setState({ folder })}
-                filter={AutoComplete.fuzzyFilter}
-                dataSource={directories}
-                openOnFocus
-              />
-            </MuiThemeProvider>
-          </label>
-          <input
-            placeholder={imgData.file}
-            value={newName}
-            onChange={(e) => this.setState({ newName: e.target.value })}
-            onKeyUp={this.fileCheck}/>
-          <button
-            className='btn'
-            onClick={this.saveFile}
-            disabled={errorMessage || !newName && !folder}
-            >Save
-          </button>
-        </div>
-        {dispErrorMessage()}
-
-
-        <div className='canvas-container'>
-          <img id='screenshot'
-               src={imgData.filePath}
-               width={imgData.width}
-               height={imgData.height}/>
-          {/* <ReactPaint {...props} /> */}
-        </div>
-
+        {displayApp()}
       </div>
     )
   }
