+++ styles/index.scss
@@ -1,16 +1,8 @@
-$sick-color: peru;
+$sick-color: #FFF;
 
 body {
   background-color: $sick-color;
   font-family: sans-serif;
-}
-
-h1 {
-  color: white;
-}
-
-canvas {
-  background-color: white;
 }
 
 .canvas-container {
